+++ src/main/java/org/zalando/nakadi/controller/EventPublishingController.java
@@ -77,8 +77,8 @@ public class EventPublishingController {
                         Problem.valueOf(Response.Status.FORBIDDEN, "Application or event type is blocked"), request);
             }
 
-            final ResponseEntity response = postEventInternal(eventTypeName, eventsAsString,
-                    request, eventTypeMetrics, client);
+            final ResponseEntity response = postEventInternal(
+                    eventTypeName, eventsAsString, request, eventTypeMetrics, client);
             eventTypeMetrics.incrementResponseCount(response.getStatusCode().value());
             return response;
         } catch (final RuntimeException ex) {
@@ -95,7 +95,7 @@ public class EventPublishingController {
             throws AccessDeniedException, ServiceTemporarilyUnavailableException {
         final long startingNanos = System.nanoTime();
         try {
-            final EventPublishResult result = publisher.publish(eventsAsString, eventTypeName, client);
+            final EventPublishResult result = publisher.publish(eventsAsString, eventTypeName);
 
             final int eventCount = result.getResponses().size();
             final int totalSizeBytes = eventsAsString.getBytes(Charsets.UTF_8).length;