+++ src/test/java/org/zalando/nakadi/domain/BatchFactoryTest.java
@@ -100,6 +100,24 @@ public class BatchFactoryTest {
     }
 
     @Test
+    public void testNoCommaBetweenEvents1() {
+        final String events = "[{\"name\":\"MyEvent\"}        {\"name\":\"MyOtherEvent\"}]";
+        final List<BatchItem> batch = BatchFactory.from(events);
+        assertEquals(2, batch.size());
+        assertEquals("{\"name\":\"MyEvent\"}", batch.get(0).getEvent().toString());
+        assertEquals("{\"name\":\"MyOtherEvent\"}", batch.get(1).getEvent().toString());
+    }
+
+    @Test
+    public void testNoCommaBetweenEvents2() {
+        final String events = "[{\"name\":\"MyEvent\"}{\"name\":\"MyOtherEvent\"}]";
+        final List<BatchItem> batch = BatchFactory.from(events);
+        assertEquals(2, batch.size());
+        assertEquals("{\"name\":\"MyEvent\"}", batch.get(0).getEvent().toString());
+        assertEquals("{\"name\":\"MyOtherEvent\"}", batch.get(1).getEvent().toString());
+    }
+
+    @Test
     public void testEmptyCharactersAroundArray() {
         final String events = "\t [{\"name\":\"MyEvent\"},{\"name\":\"MyOtherEvent\"}]\n\n";
         final List<BatchItem> batch = BatchFactory.from(events);