+++ src/test/java/org/zalando/nakadi/utils/TestUtils.java
@@ -7,6 +7,7 @@ import org.apache.commons.io.IOUtils;
 import org.echocat.jomon.runtime.concurrent.RetryForSpecifiedTimeStrategy;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
+import org.json.JSONArray;
 import org.json.JSONObject;
 import org.mockito.ArgumentCaptor;
 import org.springframework.http.converter.StringHttpMessageConverter;
@@ -15,6 +16,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.validation.Errors;
 import org.springframework.validation.FieldError;
 import org.zalando.nakadi.config.JsonConfig;
+import org.zalando.nakadi.domain.BatchFactory;
 import org.zalando.nakadi.domain.BatchItem;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.Storage;
@@ -183,7 +185,7 @@ public class TestUtils {
 
     @SuppressWarnings("unchecked")
     public static void waitFor(final Runnable runnable, final long timeoutMs, final int intervalMs,
-                               final Class<? extends  Throwable>... additionalException) {
+                               final Class<? extends Throwable>... additionalException) {
         final List<Class<? extends Throwable>> leadToRetry =
                 Stream.concat(Stream.of(additionalException), Stream.of(AssertionError.class)).collect(toList());
         executeWithRetry(
@@ -194,11 +196,11 @@ public class TestUtils {
     }
 
     public static BatchItem createBatchItem(final JSONObject event) {
-        return new BatchItem(event.toString());
+        return BatchFactory.from(new JSONArray().put(event).toString()).get(0);
     }
 
     public static BatchItem createBatchItem(final String event) {
-        return new BatchItem(event);
+        return BatchFactory.from("[" + event + "]").get(0);
     }
 
     public static DateTime randomDate() {